Start work on "Rooms"
Rooms is how we consolidate both the Chat experience and also all the
other parts of the site (Discussions, Bounties, Projects, Assets and
also Decisions).

It’s kind of amazing that this is possible to do in a progressive way.
We can do this because of the constraint that the “room” has the
discussion and that’s separate from the “object”.

Currently it’s still a WIP (not publicly accessible) as there’s a few
bugs and I haven’t done Projects or Assets yet. There’s also multiple
different ways to render timelines but they can be consolidated later.