Tip bounty authors 5% of the winnings
Apologies for the big commit. This includes a change on how minting
works as the old system was too tied to AutoTipContracts.

Now when coins are minted they flow into a wallet for the work done.
Then a job runs which divies out the coins based on the contracts on
that work. This would include auto tip contracts and also a contract
with the bounty author.