Wouldn't normally do this but it fixes a migration
Since the updated schema isn't immediately available after adding a
column during a migraiton, this worked locally, but not when run in a
single series of migrations. To fix this I edited the migration -- not
something I'm proud of, but should be ok in this instance as I'm not
changing the schema, just data.