Bring the bounty creation modal in chat up to snuff with the other bounty creation modals
We weren't using the React component in chat, so users couldn't
set the initial valuation. It also meant that we needed to change
behavior in two places any time we changed behavior in one of them.
Now chat uses the same modal that's used throughout the site, so we
should be in the clear.