Fix marking all news feed notifications as read
We were originally passing a story as a property of a subcomponent,
Entry. But properties are immutable, so even when a story was marked as
read on the server and updated accordingly in the NewsFeed state,
changes to the object would not propagate to the Entry subcomponent. I
moved the entirety of the Entry component's functionality to the
DropdownNewsFeed component.

TODO: Test/debug the FullPageNewsFeed as well.