Fix discover/updates with some caching improvements
Cuts response time for the serialized payload with a warm cache
from ~1600ms to ~1000ms, which is still not blazing fast, but we're now
using fragment caching with JBuilder so that one update doesn't blow the
entire cache. We also get users to the more quickly by serializing the
data in a separate call -- we don't want to underestimate perceived
speed gains there.