Enable uploading comment assets as product assets
This feature touched an absurd amount of the code base for something
that was seemingly so simple. Basically, when a comment is uploaded it
now sets an array of associated attachment IDs. Then, if a user clicks
"Save to product assets" on that comment, these assets are assigned to
the product -- there's no additional upload step, just a bit of
duplication in the database.

It also, sadly, means that moving away from Events is going to be
slightly more difficult, but probably still manageable.