Fix the bug where the Markdown text for asset uploads was getting clobbered
Basically, React was trying to keep track of the efficient shadow DOM,
but DropzoneView was updating the actual DOM under React's feet. This
caused conflict.

Also add duplication guards, the ability to upload individual assets

TODO: efficiently check whether an asset exists for a product
when the page is rendered