Fix issue where create bounty button was disabled when coming from the dashboard
Creating bounties from the dashboard is still broken, but it's a
slightly different issue there, and fixing it requires changing how the
bounty valuation slider gets its data. Basically, we should load that
data in lazily and provide the bounty valuation slider with the minimum
data set it needs to render.