copy edits on homepage
i couldn't help myself... the first is what we call a typesetting issue with a "widowed" or "orphaned" word (in this case "love") that's all by itself on its own line. It's just better to not have one word alone so adding the word "use" makes this look better.

the second is the use of "Internet" rather than "Information" which is typically used in conjunction with "age" (and it should be capitalized). "Information Age" is what is more commonly used.

next... I've made some small edits for flow and readability, oxford commas, consistent capitalization and word counts... etc... etc... ... these are obviously up for debate.